libratbag-hidraw: add functions for hidraw read w/ filter
Currently, we only have functions that will read and timeout after 1s if
no packets were received. The issue with this is that a lot of times we
may catch unrelated packets, such as mouse or keyboard changes.
To solve this, this patch adds a new set of functions suffixed with
_filter, that will keep reading packets until a valid packet if found.
Users define a filter function and pass it as an argument.

Signed-off-by: Filipe Laíns <lains@riseup.net>